A new function to check the visibility of a drawable
A new function has been integrated in the drawable object in order to check the visibility. This modification is part of a small change in the way the content of a camera is drawn: first the objects' coordinates (all of them) are normalized and then, into another loop, the system draws the drawable objects.